+++ extensions/src/extensions/isSelectedAnyOnPage.ts
@@ -7,6 +7,11 @@ export function isSelectedAnyOnPage<TItem>(): GridFunction<boolean>
 {
     return grid =>
     {
+        if(!grid.initializedSignal())
+        {
+            return false;
+        }
+
         const dataLoader = grid.getPlugin<DataLoader<DataResponse<TItem>>>(GridPluginType.DataLoader);
         const rowSelector = grid.getPlugin(GridPluginType.RowSelector);