+++ extensions/src/extensions/resetSelection.ts
@@ -7,6 +7,11 @@ export function resetSelection(): GridAction
 {
     return grid =>
     {
+        if(!grid.initializedSignal())
+        {
+            return;
+        }
+
         const rowSelector = grid.getPlugin<RowSelector>(GridPluginType.RowSelector);
 
         rowSelector.resetSelection();