+++ extensions/src/extensions/selectAllOnPage.ts
@@ -9,6 +9,11 @@ export function selectAllOnPage<TItem>(select: boolean = true, predicate: (item:
 {
     return grid =>
     {
+        if(!grid.initializedSignal())
+        {
+            return;
+        }
+
         const dataLoader = grid.getPlugin<DataLoader<DataResponse<TItem>>>(GridPluginType.DataLoader);
         const rowSelector = grid.getPlugin<RowSelector>(GridPluginType.RowSelector);